chore: move Execute to a separate package to be exposed and used in tests
